test(PlayersTest): 게임 시작 시 두 장의 카드 받는 테스트 추가, 승패 결정 테스트 이동
