+++ src/test/java/blackjack/domain/DealerTest.java
@@ -49,33 +49,4 @@ class DealerTest {
 
         assertThat(player1.getCards().getCards().size()).isEqualTo(1);
     }
-
-    @Test
-    @DisplayName("최종 승패를 결정한다.")
-    void decideResult() {
-        Player player3 = new Player("3");
-        dealer = new Dealer(new Players(List.of(player1, player2, player3)));
-
-        dealer.receiveCard(new Card(Number.THREE, Pattern.DIAMOND));
-        dealer.receiveCard(new Card(Number.NINE, Pattern.CLUB));
-        dealer.receiveCard(new Card(Number.EIGHT, Pattern.DIAMOND));
-
-        player1.receiveCard(new Card(Number.TWO, Pattern.HEART));
-        player1.receiveCard(new Card(Number.EIGHT, Pattern.SPADE));
-        player1.receiveCard(new Card(Number.ACE, Pattern.CLUB));
-
-        player2.receiveCard(new Card(Number.SEVEN, Pattern.CLUB));
-        player2.receiveCard(new Card(Number.K, Pattern.SPADE));
-
-        player3.receiveCard(new Card(Number.THREE, Pattern.HEART));
-        player3.receiveCard(new Card(Number.NINE, Pattern.SPADE));
-        player3.receiveCard(new Card(Number.EIGHT, Pattern.CLUB));
-
-        Map<Player, Result> result = dealer.decideResult();
-        assertAll(
-                () -> assertThat(result.get(player1)).isEqualTo(Result.WIN),
-                () -> assertThat(result.get(player2)).isEqualTo(Result.LOSE),
-                () -> assertThat(result.get(player3)).isEqualTo(Result.DRAW)
-        );
-    }
 }
